requirements.txt
attrs==23.1.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.2.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==23.1
pexpect==4.8.0
ptyprocess==0.7.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
sqlparser==0.0.9
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
websocket-client==1.6.1
wsproto==1.2.0
psycopg2-binary==2.9.9
